+++ cli/src/custom/help.ts
@@ -11,11 +11,10 @@ import { Command, Help, Interfaces } from "@oclif/core";
 import { getHeader } from "./header.js";
 import chalk from "chalk";
 
-export default class CustomHelp extends Help {
-  private target_pad = 15;
-  private pre_pad = 0;
-  private post_pad = 0;
+const FIRST_PAD = 2;
+const SECOND_PAD = 20;
 
+export default class CustomHelp extends Help {
   formatHeader(): string {
     return getHeader(this.config.version);
   }
@@ -33,25 +32,10 @@ export default class CustomHelp extends Help {
 
       const rawName = command.id.includes(":") ? command.id.split(":")[1] : command.id;
       const name = chalk.bold.blueBright(rawName);
-      const prePadding = " ".repeat(Math.max(1, this.pre_pad - rawName.length));
-      const args =
-        Object.keys(command.args).length > 0
-          ? Object.entries(command.args)
-              .map((v) => {
-                if (!v[1].hidden && v[1].required) {
-                  if (v[1].description && v[1].description.indexOf("-|-") > 0) {
-                    return v[1].description.split("-|-")[0];
-                  }
-                }
-                return "";
-              })
-              .join(" ")
-          : "";
-      const postPadding = " ".repeat(Math.max(1, this.post_pad - args.length));
       const description = command.description!;
       const aliases = command.aliases.length > 0 ? chalk.dim(" (" + command.aliases.join("/") + ")") : "";
 
-      out += "  " + name + prePadding + chalk.dim(args) + postPadding + description + aliases + "\n";
+      out += getMessageWithPad(name, description + aliases, FIRST_PAD, SECOND_PAD, rawName.length) + "\n";
     }
     return out.trim();
   }
@@ -70,27 +54,15 @@ export default class CustomHelp extends Help {
 
     for (const topic of topics) {
       if (topic.hidden) continue;
-      out += "  " + chalk.bold.blue(topic.name) + " ".repeat(Math.max(1, this.pre_pad + this.post_pad - topic.name.length)) + topic.description + "\n";
+
+      out += getMessageWithPad(chalk.bold.blue(topic.name), topic.description || "", FIRST_PAD, SECOND_PAD, topic.name.length) + "\n";
     }
     return out.trim();
   }
 
-  // formatFlags(topics: Interfaces.Topic[]): string {
-  //   let out = "";
-  //   if (topics.find((v) => !v.hidden)) out += chalk.bold("Flags:") + "\n";
-  //   else return out;
-
-  //   for (const topic of topics) {
-  //     if (topic.hidden) continue;
-
-  //     const fullName = topic.shortName ? `${topic.shortName}, ${topic.name}` : topic.name;
-  //     out += "  " + chalk.bold.blue(fullName) + " ".repeat(Math.max(1, this.pre_pad + this.post_pad - fullName.length)) + topic.description + "\n";
-  //   }
-  //   return out.trim();
-  // }
-
   formatFooter(): string {
     let out = "";
+
     const links = [
       { name: "Docs", url: "https://docs.hypermode.com/modus" },
       { name: "GitHub", url: "https://github.com/hypermodeinc/modus" },
@@ -98,7 +70,7 @@ export default class CustomHelp extends Help {
     ];
 
     for (const link of links) {
-      out += `${link.name}: ${" ".repeat(Math.max(1, this.pre_pad + this.post_pad - link.name.length))}${link.url}\n`;
+      out += getMessageWithPad(link.name, link.url, 0, SECOND_PAD) + "\n";
     }
 
     out += "\n";
@@ -125,26 +97,6 @@ export default class CustomHelp extends Help {
       rootCommands = rootCommands.filter((c) => !c.id.includes(":"));
     }
 
-    for (const command of rootCommands) {
-      if (command.id.length > this.pre_pad) this.pre_pad = command.id.length;
-      const args =
-        Object.keys(command.args).length > 0
-          ? Object.entries(command.args)
-              .map((v) => {
-                if (!v[1].hidden && v[1].required) {
-                  if (v[1].description && v[1].description.indexOf("-|-") > 0) {
-                    return v[1].description.split("-|-")[0];
-                  }
-                }
-                return "";
-              })
-              .join(" ")
-          : "";
-      if (args.length > this.post_pad) this.post_pad = args.length;
-    }
-    this.post_pad = 6 + this.pre_pad + this.post_pad > this.target_pad ? 6 + this.pre_pad + this.post_pad - this.target_pad : this.target_pad - this.pre_pad;
-    this.pre_pad += 2;
-
     if (rootCommands.length > 0) {
       rootCommands = rootCommands.filter((c) => c.id);
       this.log(this.formatCommands(rootCommands));
@@ -156,22 +108,36 @@ export default class CustomHelp extends Help {
       this.log();
     }
 
-    // const globalFlagTopics: Interfaces.Topic[] = [
-    //   {
-    //     name: "--help",
-    //     description: "Show help message",
-    //     shortName: "-h",
-    //   },
-    //   {
-    //     name: "--version",
-    //     description: "Show Modus version",
-    //     shortName: "-v",
-    //   },
-    // ];
-    // this.log(this.formatFlags(globalFlagTopics));
-    // this.log();
+    const globalFlagTopics: Interfaces.Topic[] = [
+      {
+        name: "--help",
+        description: "Show help message",
+        shortName: "-h",
+      },
+      {
+        name: "--version",
+        description: "Show Modus version",
+        shortName: "-v",
+      },
+    ];
+    this.log(this.formatRootFlags(globalFlagTopics));
+    this.log();
 
     this.log(this.formatFooter());
+  }
+
+  formatRootFlags(topics: Interfaces.Topic[]): string {
+    let out = "";
+    if (topics.find((v) => !v.hidden)) out += chalk.bold("Flags:") + "\n";
+    else return out;
+
+    for (const topic of topics) {
+      if (topic.hidden) continue;
+
+      const fullName = topic.shortName ? `${topic.shortName}, ${topic.name}` : topic.name;
+      out += getMessageWithPad(chalk.bold.blue(fullName), topic.description || "", FIRST_PAD, SECOND_PAD, fullName.length) + "\n";
+    }
+    return out.trim();
   }
 
   async showTopicHelp(topic: Interfaces.Topic) {
@@ -179,25 +145,7 @@ export default class CustomHelp extends Help {
 
     const { name } = topic;
     const commands = this.sortedCommands.filter((c) => c.id.startsWith(name + ":"));
-    for (const command of commands) {
-      if (command.id.split(":")[1].length > this.pre_pad) this.pre_pad = command.id.split(":")[1].length;
-      const args =
-        Object.keys(command.args).length > 0
-          ? Object.entries(command.args)
-              .map((v) => {
-                if (!v[1].hidden && v[1].required) {
-                  if (v[1].description && v[1].description.indexOf("-|-") > 0) {
-                    return v[1].description.split("-|-")[0];
-                  }
-                }
-                return "";
-              })
-              .join(" ")
-          : "";
-      if (args.length > this.post_pad) this.post_pad = args.length;
-    }
-    this.post_pad = 6 + this.pre_pad + this.post_pad > this.target_pad ? 6 + this.pre_pad + this.post_pad - this.target_pad : this.target_pad - this.pre_pad;
-    this.pre_pad += 2;
+
     const state = this.config.pjson?.oclif?.state;
     if (state) this.log(`This topic is in ${state}.\n`);
     this.log(this.formatTopic(topic));
@@ -212,7 +160,6 @@ export default class CustomHelp extends Help {
   async showCommandHelp(command: Command.Loadable): Promise<void> {
     this.log(this.formatHeader());
 
-    const margin = 20;
     const name = command.id.replaceAll(":", " ");
     const args = Object.keys(command.args);
     const flags = Object.keys(command.flags);
@@ -227,16 +174,20 @@ export default class CustomHelp extends Help {
       this.log(chalk.bold("Args:"));
       for (let arg of Object.values(command.args)) {
         let usage = "";
-        let desc = arg.description;
+        let desc = arg.description || "";
         if (arg.description?.includes("-|-")) {
           usage = arg.description.split("-|-")[0];
           desc = arg.description.split("-|-")[1];
         }
-        let defaultValue = "";
         if (arg.default) {
-          defaultValue = chalk.dim(` (default: '${arg.default}')`);
+          desc += chalk.dim(` (default: ${arg.default})`);
         }
-        this.log("  " + chalk.bold.blueBright(arg.name) + " ".repeat(Math.max(1, margin + 2 - arg.name.length)) + desc + defaultValue);
+
+        if (arg.options) {
+          desc += chalk.dim(` (options: ${arg.options.join(", ")})`);
+        }
+
+        this.log(getMessageWithPad(chalk.bold.blueBright(arg.name), desc, FIRST_PAD, SECOND_PAD, arg.name.length));
       }
       this.log();
     }
@@ -247,11 +198,33 @@ export default class CustomHelp extends Help {
         if (flag.hidden) continue;
 
         const flagOptions = flag.char ? `-${flag.char}, --${flag.name}` : `--${flag.name}`;
-        this.log("  " + chalk.bold.blueBright(flagOptions) + " ".repeat(margin + 2 - flagOptions.length) + flag.description);
+
+        let desc = flag.description || "";
+        // The interface doesn't have `options`, but `options` is a valid property of a flag
+        if ("options" in flag && flag.options) {
+          desc += chalk.dim(` (${flag.options.join(", ")})`);
+        }
+
+        if (flag) this.log(getMessageWithPad(chalk.bold.blueBright(flagOptions), desc, FIRST_PAD, SECOND_PAD, flagOptions.length));
       }
       this.log();
     }
 
     this.log(this.formatFooter());
   }
+}
+
+/**
+ * Used for printing help messages like:
+ *   build               Build a Modus app
+ *   dev                 Run a Modus app locally for development
+ *
+ * @param firstMessageLength The message can be ANSI escaped and have a length different from the actual string length
+ */
+function getMessageWithPad(formattedFirstMsg: string, formattedSecondMsg: string, firstPad: number, secondPad: number, firstMessageLength?: number): string {
+  const padLeft = " ".repeat(firstPad);
+  const actualFirstMsgLength = firstMessageLength ?? formattedFirstMsg.length;
+  const padMiddle = " ".repeat(secondPad - firstPad - actualFirstMsgLength);
+
+  return padLeft + formattedFirstMsg + padMiddle + formattedSecondMsg;
 }